Renamed old sampler class.
